+++ include/seqan3/alphabet.hpp
@@ -48,7 +48,7 @@
 
 #include <seqan3/alphabet/concept.hpp>
 #include <seqan3/alphabet/nucleotide/all.hpp>
-#include <seqan3/alphabet/quality.hpp>
+#include <seqan3/alphabet/quality/all.hpp>
 #include <seqan3/alphabet/aminoacid/all.hpp>
 #include <seqan3/alphabet/gap/gap.hpp>
 #include <seqan3/alphabet/composition/all.hpp>
